Export a single graphql schema (#646)
