+++ packages/kiosk/.prettierignore
@@ -1,2 +1,3 @@
 dist
 CHANGELOG.md
+test/e2e/data
