+++ packages/typescript/graphql/schema/package.json
@@ -1,0 +1,6 @@
+{
+	"private": true,
+	"import": "../../dist/esm/graphql/schema/index.js",
+	"main": "../../dist/cjs/graphql/schema/index.js",
+	"sideEffects": false
+}
