--- packages/typescript/graphql/schemas/2024.4/package.json
@@ -1,6 +1,0 @@
-{
-	"private": true,
-	"import": "../../../dist/esm/graphql/schemas/2024.4/index.js",
-	"main": "../../../dist/cjs/graphql/schemas/2024.4/index.js",
-	"sideEffects": false
-}
