--- packages/typescript/graphql/schemas/latest/package.json
@@ -1,6 +1,0 @@
-{
-	"private": true,
-	"import": "../../../dist/esm/graphql/schemas/latest/index.js",
-	"main": "../../../dist/cjs/graphql/schemas/latest/index.js",
-	"sideEffects": false
-}
