+++ packages/typescript/package.json
@@ -101,17 +101,9 @@
 			"import": "./dist/esm/zklogin/index.js",
 			"require": "./dist/cjs/zklogin/index.js"
 		},
-		"./graphql/schemas/2024.1": {
-			"import": "./dist/esm/graphql/schemas/2024.1/index.js",
-			"require": "./dist/cjs/graphql/schemas/2024.1/index.js"
-		},
-		"./graphql/schemas/2024.4": {
-			"import": "./dist/esm/graphql/schemas/2024.4/index.js",
-			"require": "./dist/cjs/graphql/schemas/2024.4/index.js"
-		},
-		"./graphql/schemas/latest": {
-			"import": "./dist/esm/graphql/schemas/latest/index.js",
-			"require": "./dist/cjs/graphql/schemas/latest/index.js"
+		"./graphql/schema": {
+			"import": "./dist/esm/graphql/schema/index.js",
+			"require": "./dist/cjs/graphql/schema/index.js"
 		}
 	},
 	"scripts": {
@@ -137,8 +129,7 @@
 		"eslint:fix": "pnpm run eslint:check --fix",
 		"lint": "pnpm run eslint:check && pnpm run prettier:check",
 		"lint:fix": "pnpm run eslint:fix && pnpm run prettier:fix",
-		"update-schemas": "pnpm tsx scripts/update-graphql-schemas.ts",
-		"generate-schema": "gql.tada generate-output"
+		"update-graphql-schema": "pnpm tsx scripts/update-graphql-schema.ts"
 	},
 	"bugs": {
 		"url": "https://github.com/MystenLabs/ts-sdks/issues/new"
