+++ packages/typescript/src/graphql/schema/index.ts
@@ -3,10 +3,10 @@
 
 import { initGraphQLTada } from 'gql.tada';
 
-import type { introspection } from '../../generated/2024.4/tada-env.js';
-import type { CustomScalars } from '../../types.js';
+import type { introspection } from '../generated/tada-env.js';
+import type { CustomScalars } from '../types.js';
 
-export type * from '../../types.js';
+export type * from '../types.js';
 
 export type { FragmentOf, ResultOf, VariablesOf, TadaDocumentNode } from 'gql.tada';
 export { readFragment, maskFragments } from 'gql.tada';
