--- packages/typescript/src/graphql/schemas/latest/index.ts
@@ -1,17 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-import { initGraphQLTada } from 'gql.tada';
-
-import type { introspection } from '../../generated/latest/tada-env.js';
-import type { CustomScalars } from '../../types.js';
-
-export type * from '../../types.js';
-
-export type { FragmentOf, ResultOf, VariablesOf, TadaDocumentNode } from 'gql.tada';
-export { readFragment, maskFragments } from 'gql.tada';
-
-export const graphql = initGraphQLTada<{
-	introspection: typeof introspection;
-	scalars: CustomScalars;
-}>();
